backend/requirements.txt
fastapi
uvicorn
structlog
pydantic
pydantic-settings
requests
beautifulsoup4
tenacity
psycopg2-binary
mcp
supabase
feedparser
